small fix for backups
git-svn-id: https://plugins.svn.wordpress.org/mainwp-child/trunk@798762 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ class/MainWPHelper.class.php
@@ -426,15 +426,18 @@ class MainWPHelper
     public static function delete_dir( $dir ) {
         $nodes = glob($dir . '*');
 
-        foreach ($nodes as $node)
+        if (is_array($nodes))
         {
-            if (is_dir($node))
+            foreach ($nodes as $node)
             {
-                self::delete_dir($node . DIRECTORY_SEPARATOR);
-            }
-            else
-            {
-                @unlink($node);
+                if (is_dir($node))
+                {
+                    self::delete_dir($node . DIRECTORY_SEPARATOR);
+                }
+                else
+                {
+                    @unlink($node);
+                }
             }
         }
         @rmdir($dir);
